support decimals in api request

+++ api/v1/matic/transfer/transfer_erc20/types.go
@@ -12,10 +12,10 @@ type TransferPayload struct {
 }
 
 type TransferRequestSalt struct {
-	WalletAddress   string `json:"walletAddress" binding:"required"`
-	Mnemonic        string `json:"mnemonic" binding:"required"`
-	To              string `json:"to" binding:"required"`
-	Amount          int64  `json:"amount" binding:"required"`
-	ContractAddress string `json:"contractAddress" binding:"required"`
-	Salt            string `json:"salt" binding:"required"`
+	WalletAddress   string  `json:"walletAddress" binding:"required"`
+	Mnemonic        string  `json:"mnemonic" binding:"required"`
+	To              string  `json:"to" binding:"required"`
+	Amount          float64 `json:"amount" binding:"required"`
+	ContractAddress string  `json:"contractAddress" binding:"required"`
+	Salt            string  `json:"salt" binding:"required"`
 }
