+++ api/v1/matic/transfer/transfer_native/types.go
@@ -11,9 +11,9 @@ type TransferPayload struct {
 }
 
 type TransferRequestSalt struct {
-	WalletAddress string `json:"walletAddress" binding:"required"`
-	Mnemonic      string `json:"mnemonic" binding:"required"`
-	To            string `json:"to" binding:"required"`
-	Amount        int64  `json:"amount" binding:"required"`
-	Salt          string `json:"salt" binding:"required"`
+	WalletAddress string  `json:"walletAddress" binding:"required"`
+	Mnemonic      string  `json:"mnemonic" binding:"required"`
+	To            string  `json:"to" binding:"required"`
+	Amount        float64 `json:"amount" binding:"required"`
+	Salt          string  `json:"salt" binding:"required"`
 }
